user cart done, probebly

+++ database/migrations/2022_12_02_151734_create_store_product_discounts_table.php
@@ -17,8 +17,8 @@ class CreateStoreProductDiscountsTable extends Migration
             $table->id();
 
             $table->string('discount_type');
-            $table->unsignedInteger('discount_value');
-            $table->unsignedInteger('final_price');
+            $table->unsignedBigInteger('discount_value');
+            $table->unsignedBigInteger('final_price');
 
             $table->unsignedBigInteger('product_id');
